+++ java/client/src/org/openqa/selenium/internal/seleniumemulation/GetElementWidth.java
@@ -31,6 +31,6 @@ public class GetElementWidth extends SeleneseCommand<Number> {
   @Override
   protected Number handleSeleneseCommand(WebDriver driver, String locator, String ignored) {
     Dimension size = ((RenderedWebElement) finder.findElement(driver, locator)).getSize();
-    return (int) size.getWidth();
+    return size.getWidth();
   }
 }
